tests/wpt/tests/html/dom/render-blocking/element-render-blocking-020.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Rel attribute changes in the body to not apply</title>

<link id=link rel=expect href="#last" blocking="render">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "changing rel to non-expect in the body makes it non blocking");
</script>
</head>
<body>
<script>
link.rel = "stylesheet";
</script>
  <div id="first"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="second"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="last"></div>
</body>
